@@ -2,10 +2,10 @@ name: flutter_auth0
 description: Flutter toolkit for Auth0 API and web pkce flow authentication
 author: Denny Segura <dev.dennysegura@gmail.com>
 homepage: https://github.com/devdennysegura/flutter-auth0
-version: 0.6.0
+version: 0.6.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=2.0.0 <3.0.0"
 
 dependencies:
